requirements.txt
bitly-api
python-twitter
boto
docopt
flask
flask-flatpages
frozen-flask
flask-assets
cssmin
stylus
